Feature - Ghost Strategy - Refactored common PacMan and Ghost behaviour into a parent MovingEntity abstract class
